Converted UCUbuntuShape::sourceOpacity from float to qreal.

+++ modules/Ubuntu/Components/plugin/ucubuntushape.cpp
@@ -399,12 +399,12 @@ void UCUbuntuShape::setSource(const QVariant& source)
 
     \note Setting this disables support for the deprecated \l image property.
 */
-void UCUbuntuShape::setSourceOpacity(float sourceOpacity)
+void UCUbuntuShape::setSourceOpacity(qreal sourceOpacity)
 {
     dropImageSupport();
 
     const quint8 sourceOpacityPacked =
-        qMax(0.0f, qMin(1.0f, sourceOpacity)) * static_cast<float>(0xff);
+        qMax(0.0, qMin(1.0, sourceOpacity)) * static_cast<qreal>(0xff);
     if (m_sourceOpacity != sourceOpacityPacked) {
         m_sourceOpacity = sourceOpacityPacked;
         update();
